+++ metadata.rb
@@ -3,16 +3,13 @@ maintainer_email "a.rosenhagen@ffuenf.de"
 license          "Apache 2.0"
 description      "installs/configures dop_nginx"
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version          "1.0.5"
+name             "dop_nginx"
+version          "1.0.6"
 
 %w{ debian }.each do |os|
   supports os
 end
 
-%w{ 
-  dop_base
-  nginx
-  logrotate
-}.each do |ressource|
+%w{ nginx logrotate }.each do |ressource|
   depends ressource
 end