+++ test/integration/default/serverspec/geoip_spec.rb
@@ -1,0 +1,14 @@
+require 'serverspec'
+
+include Serverspec::Helper::Exec
+include Serverspec::Helper::DetectOS
+
+RSpec.configure do |c|
+  c.before :all do
+    c.path = '/sbin:/usr/sbin'
+  end
+end
+
+describe cron do
+  it { should have_entry '# Chef Name: run geoip update' }
+end